feat(image_processing): image processing init and fixes

+++ endpoints_detector.py
@@ -12,10 +12,10 @@ import cm_utils
 import random
 from datetime import datetime
 
-from lib.detector import Detector
+from library.detector import Detector
 
 
-detector = Detector("lib/model.pt")
+detector = Detector("library/model.pt")
 
 
 def generate_mock_data():
@@ -130,9 +130,9 @@ def set_detector_config(detector_id):
 @app.route("/detector/<detector_id>", methods=["DELETE"])
 def delete_detector(detector_id):
     try:
-        # user_data = cm_utils.auth_token()
-        # if user_data is None:
-        #     return error_response("/add_detector", "no user signed in")
+        user_data = cm_utils.auth_token()
+        if user_data is None:
+            return error_response("/add_detector", "no user signed in")
 
         mongo.logs.delete_one({"detector_id": detector_id})
